requirements.txt
# Core dependencies
numpy>=1.24.0
pyserial>=3.5

# Supabase
supabase>=2.0.0

# Firebase
firebase-admin>=6.0.0

# Console UI
rich>=13.0.0

# Async
asyncio-mqtt>=0.16.0  # Optional: MQTT support

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
